requirements.txt
This explains:

✅ What your project is

✅ What you are using

✅ What you changed from base paper

✅ What you added

✅ Folder structure

✅ Why each folder is used

✅ What code should go in EACH FILE

I’ll keep it simple, clean, and unambiguous.

🔷 OVERALL PROJECT DESCRIPTION (FINAL)
🔹 Project Title

YOLOv8-Based Multi-Class Brain Tumor Detection and Localization from MRI Images with Web Application Support

1️⃣ What is my project?

This project aims to detect, classify, and localize brain tumors from MRI images using YOLOv8, an advanced deep-learning object detection model.

The system:

Accepts MRI images as input

Detects tumor regions

Classifies tumor type (multi-class)

Displays bounding boxes with confidence scores

Provides a web-based interface using Streamlit, inspired by a smart healthcare system

2️⃣ What BASE PAPER idea am I using?

From the base paper, I use:

MRI-based brain tumor diagnosis

Deep learning for medical imaging

Smart healthcare workflow (upload → process → result)

3️⃣ What did I CHANGE from the base paper?
🔁 CNN → YOLOv8

Base paper uses CNN for classification

I replaced CNN with YOLOv8 for:

Object detection

Tumor localization

Multi-class classification

🔁 Binary → Multi-Class

Instead of tumor / no tumor

I classify:

Glioma

Meningioma

Pituitary tumor

🔁 No CNN Baseline

I do NOT use CNN baseline

Reason: CNN cannot localize tumors

YOLOv8 alone satisfies project objective

4️⃣ What did I ADD to make the project stronger?

YOLOv8 object detection

Multi-class tumor classification

Bounding box visualization

Modular project structure

Streamlit-based web application

Clean separation of data, model, app, and results

5️⃣ Technologies Used

Language: Python

Model: YOLOv8 (Ultralytics)

Domain: Medical Imaging, Deep Learning

Frameworks/Libraries:

ultralytics

PyTorch

OpenCV

NumPy

Streamlit

Dataset: Brain MRI images (YOLO-formatted)

6️⃣ FINAL PROJECT FOLDER STRUCTURE
Brain-Tumor-Detection-YOLO/
│
├── app/
│   └── app.py
│
├── data/
│   ├── raw_images/
│   │   ├── glioma/
│   │   ├── meningioma/
│   │   ├── pituitary/
│   │   └── no_tumor/
│   │
│   ├── yolo_dataset/
│   │   ├── images/
│   │   │   ├── train/
│   │   │   ├── val/
│   │   │   └── test/
│   │   │
│   │   ├── labels/
│   │   │   ├── train/
│   │   │   ├── val/
│   │   │   └── test/
│   │
│   └── data.yaml
│
├── preprocessing/
│   ├── resize_images.py
│   ├── normalize_images.py
│   ├── augmentation.py
│   └── convert_to_yolo_format.py
│
├── model/
│   ├── train_yolo.py
│   └── detect_yolo.py
│
├── evaluation/
│   ├── metrics.py
│   └── results_analysis.py
│
├── visualization/
│   └── sample_outputs/
│
├── results/
│   ├── training_logs/
│   ├── graphs/
│   └── final_outputs/
│
├── docs/
│   └── project_report.pdf
│
├── README.md
└── requirements.txt

7️⃣ WHAT EACH FOLDER & FILE DOES (IMPORTANT)
📁 data/

Purpose: Dataset storage

raw_images/ → original MRI images (class-wise)

yolo_dataset/ → YOLO-formatted images & labels

data.yaml → YOLO dataset configuration

📌 data.yaml should define:

Dataset path

Train/val/test folders

Class names

📁 preprocessing/

Purpose: Image preparation before training

resize_images.py → resize MRI images

normalize_images.py → normalize pixel values

augmentation.py → rotate/flip/zoom images

convert_to_yolo_format.py → convert annotations to YOLO format

📁 model/

Purpose: YOLOv8 logic

train_yolo.py

Load YOLOv8 pretrained model

Train on MRI dataset

Save best model (best.pt)

detect_yolo.py

Load trained YOLOv8 model

Perform detection on new images

Return bounding boxes, class labels, confidence

📁 app/

Purpose: Web application (Streamlit)

app.py

Upload MRI image

Call YOLOv8 detection

Display result image with bounding boxes

Show tumor type and confidence

📌 This matches the base paper smart healthcare interface

📁 evaluation/

Purpose: Model evaluation

metrics.py → calculate precision, recall, F1-score

results_analysis.py → summarize performance

📁 visualization/

Purpose: Output images

sample_outputs/ → detection result images

📁 results/

Purpose: Store outputs

training_logs/ → YOLO training logs

graphs/ → accuracy & loss graphs

final_outputs/ → final prediction images

📁 docs/

Purpose: Documentation

Project report

Architecture diagrams

Base paper reference

📄 README.md

Purpose: Project overview & instructions

Should include:

Project description

How to install dependencies

How to train YOLOv8

How to run Streamlit app

📄 requirements.txt

Purpose: Dependency list

Example:

ultralytics
torch
opencv-python
numpy
streamlit

8️⃣ END-TO-END WORKFLOW
MRI Image Upload
      ↓
Preprocessing
      ↓
YOLOv8 Model
      ↓
Tumor Detection & Localization
      ↓
Multi-Class Classification
      ↓
Bounding Box Visualization
      ↓
Web App Display (Streamlit)

9️⃣ ONE-PARAGRAPH FINAL SUMMARY (REPORT-READY)

This project implements a YOLOv8-based multi-class brain tumor detection and localization system using MRI images. Inspired by a deep learning-based smart healthcare framework, the system replaces traditional CNN classifiers with an advanced object detection model to accurately identify tumor presence, type, and location. The project is organized into modular components for data handling, preprocessing, model training, evaluation, visualization, and web deployment, enabling scalable and interpretable medical image analysis.

streamlit
ultralytics
pillow
numpy
opencv-python-headless